pyproject.toml
[project]
name = "walla-walla-course-search"
version = "0.1.0"
description = "Semantic search for Walla Walla University Computer Science courses using word2vec"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "gensim>=4.3.0",
    "numpy>=1.24.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["walla_walla_course_search"]
